Rename tests/performance_benchmarks to tests/benchmarks and update docs
- Rename tests/performance_benchmarks/ to tests/benchmarks/ to follow S-CORE naming conventions
- Rename PERFORMANCE_ANALYSIS_DRAFT.md to benchmarks_analysis.md
- Add MAIN_LOOP_SLEEP measurement caveat to benchmarks_analysis.md
- Improve deployment/BUILD.bazel platform documentation
