+++ deployment/BUILD.bazel
@@ -12,77 +12,92 @@
 # *******************************************************************************
 
 # =============================================================================
-# Simple Deployment Packages
+# Deployment packages — one build invocation targets one platform.
 #
-# Creates platform-specific deployment packages (tar.gz) containing:
-# - Executables (bins with all runtime dependencies)
-# - Configuration files (.json)
+# Pass --config to select the cross-compilation toolchain:
+#   x86_64-linux    Linux / x86_64   (developer workstation)
+#   aarch64-linux   Linux / aarch64  (e.g. Raspberry Pi 4 + Ubuntu)
+#   x86_64-qnx     QNX 8 / x86_64
+#   aarch64-qnx    QNX 8 / aarch64  (e.g. Raspberry Pi 4 + QNX 8)
 #
-# Usage:
-#   # Build arm64 packages
-#   bazel build //deployment:someipd_tar --config=arm64-linux
-#   bazel build //deployment:gatewayd_tar --config=arm64-linux
-#   bazel build //deployment:benchmarks_tar --config=arm64-linux
+# The output archive name contains the CPU identifier from the active toolchain
+# (e.g. "aarch64" for both aarch64-linux and aarch64-qnx builds).
 #
-#   # Build x86_64 packages (default)
-#   bazel build //deployment:someipd_tar
-#   bazel build //deployment:gatewayd_tar
-#   bazel build //deployment:benchmarks_tar
+# Build individual packages:
+#   bazel build //deployment:gateway_tar    --config=aarch64-linux
+#   bazel build //deployment:benchmarks_tar --config=aarch64-linux
 #
-#   # Build all
-#   bazel build //deployment:all_tars --config=arm64-linux
-#   bazel build //deployment:all_tars
+# Build all packages for a platform in one invocation:
+#   bazel build //deployment:all_tars --config=aarch64-linux
+#
+# Repeat with a different --config to produce packages for another platform.
 # =============================================================================
 
-load("@rules_pkg//pkg:tar.bzl", "pkg_tar")
-load(":defs.bzl", "names_from_toolchains")
+load("@rules_pkg//pkg:mappings.bzl", "pkg_files")
+load(":defs.bzl", "names_from_toolchains", "scoreipgw_pkg_tar")
 
 names_from_toolchains(
     name = "toolchain_vars",
 )
 
-# TODO refactor common attributes into a macro
-pkg_tar(
-    name = "someipd_tar",
+# Config files are renamed and flattened to the package root so that
+# run_gateway.sh can reference them via ${SCRIPT_DIR}/<filename>.
+pkg_files(
+    name = "gatewayd_config_files",
     srcs = [
+        "//src/gatewayd:config_file",
+        "//src/gatewayd:etc/mw_com_config.json",
+    ],
+    renames = {
+        "//src/gatewayd:config_file": "gatewayd_config.bin",
+        "//src/gatewayd:etc/mw_com_config.json": "gatewayd_mw_com_config.json",
+    },
+)
+
+pkg_files(
+    name = "someipd_config_files",
+    srcs = ["//src/someipd:etc/mw_com_config.json"],
+    renames = {
+        "//src/someipd:etc/mw_com_config.json": "someipd_mw_com_config.json",
+    },
+)
+
+# =============================================================================
+# gateway_tar — gateway daemon package
+#
+# Contains: gatewayd, someipd (with Bazel runfiles), default configs, and
+#           run_gateway.sh for easy startup on the target machine.
+#
+# The archive is named gateway_tar-<cpu>.tar.gz, where <cpu> is the CPU
+# identifier resolved by the active toolchain (e.g. gateway_tar-aarch64.tar.gz
+# when using --config=aarch64-linux or --config=aarch64-qnx).
+#
+# Deploy on the target:
+#   tar xzf gateway_tar-<cpu>.tar.gz
+#   export VSOMEIP_CONFIGURATION=/path/to/vsomeip.json
+#   ./score_someip_gateway/gateway/run_gateway.sh gatewayd   # terminal 1
+#   ./score_someip_gateway/gateway/run_gateway.sh someipd    # terminal 2
+# =============================================================================
+
+scoreipgw_pkg_tar(
+    name = "gateway_tar",
+    srcs = [
+        "run_gateway.sh",
+        ":gatewayd_config_files",
+        ":someipd_config_files",
+        "//src/gatewayd",
         "//src/someipd",
     ],
-    out = "someipd.tar.gz",
-    extension = "tar.gz",
-    include_runfiles = True,
-    package_dir = module_name(),
-    package_file_name = "someipd-{cc_cpu}.tar.gz",
-    package_variables = ":toolchain_vars",
-    visibility = ["//visibility:public"],
+    package_dir = module_name() + "/gateway",
 )
 
-pkg_tar(
-    name = "gatewayd_tar",
-    srcs = [
-        "//src/gatewayd",
-    ],
-    out = "gatewayd.tar.gz",
-    extension = "tar.gz",
-    include_runfiles = True,
-    package_dir = module_name(),
-    package_file_name = "gatewayd-{cc_cpu}.tar.gz",
-    package_variables = ":toolchain_vars",
-    visibility = ["//visibility:public"],
-)
-
-pkg_tar(
+scoreipgw_pkg_tar(
     name = "benchmarks_tar",
     srcs = [
-        "//tests/performance_benchmarks:echo_server",
-        "//tests/performance_benchmarks:ipc_benchmarks",
+        "//tests/benchmarks:echo_server",
+        "//tests/benchmarks:ipc_benchmarks",
     ],
-    out = "benchmarks.tar.gz",
-    extension = "tar.gz",
-    include_runfiles = True,
     package_dir = module_name() + "/benchmarks",
-    package_file_name = "benchmarks-{cc_cpu}.tar.gz",
-    package_variables = ":toolchain_vars",
-    visibility = ["//visibility:public"],
 )
 
 # =============================================================================
@@ -93,8 +108,7 @@ filegroup(
     name = "all_tars",
     srcs = [
         ":benchmarks_tar",
-        ":gatewayd_tar",
-        ":someipd_tar",
+        ":gateway_tar",
     ],
     visibility = ["//visibility:public"],
 )
