+++ src/someipd/BUILD.bazel
@@ -18,6 +18,8 @@ This daemon contains the SOME/IP communication stack and is QM
 load("@rules_cc//cc:cc_binary.bzl", "cc_binary")
 load("@score_communication//bazel/tools:json_schema_validator.bzl", "validate_json_schema_test")
 
+exports_files(["etc/mw_com_config.json"])
+
 # ============================================================================
 # Main Binary
 # ============================================================================
